Rebasing openstack-helm-infra
kube-state-metrics: use endpoints section and lookups to set port

This PS moves kube-state-metrics to use the endpoints section and
lookups to set the ports it serves on.

Change-Id: Icb4757a59852e508148ca9f1e682c722e40042c9

Remove Api version requirement for CronJob Curator

Curator job in Elasticsearch helm chart has a condition on
api version batch/v2alpha1. Cronjob resource is deprecated
in batch/v2alpha1 from k8s 1.8 and batch/v1beta1 is enabled
by default.

Remove the condition on API version as it is no more required.

Closes-Bug: #1753524
Change-Id: Ia296b3742e655fae508e5d4402e7f3881db31688

Fluent-logging: use endpoints section and lookups to set port

This PS moves fluent-logging to use the endpoints section and
lookups to set the port it serves on.

Change-Id: I7cbbd8d6287942eb36f70ae74872405038e523e8

Elasticsearch: use endpoints section and lookups to set port

This PS moves elasticsearch to use the endpoints section and
lookups to set the port it serves on.

Change-Id: I4a73893124b6d988cd1f885cfc3dd62abeb4ae8c

Prometheus: use endpoints section and lookups to set port

This PS moves prometheus to use the endpoints section and lookups
to set the port it serves on.

Change-Id: Ifae665e21128dd566da5a68b9904a94a68df4018

Grafana: use endpoints section and lookups to set port

This PS moves grafana to use the endpoints section and lookups to
set the port it serves on.

Change-Id: I51d4c10297f3423569539dcbf7fbecad24d5e47f

Kibana: use endpoints section and lookups to set port

This PS moves kibana to use the endpoints section and lookups to
set the port it serves on.

Change-Id: I710428f92e80faf6ac5bb444f938447248e99217

Alertmanager: use endpoints section and lookups to set port

This PS moves alertmanager to use the endpoints section and
lookups to set the ports it serves on.

Change-Id: I62108ca207f615d10d0b4385da204214b9aeae32

yaml cleanup: trim multiline strings

Change-Id: I7e8f423be2efb84f3116258beca805265ca388f7

KubeADM: force calico interface to match to same as kubelet

This PS forces calico to use the same interface as the kubelet
for inter-node communication.

Change-Id: I7d03beec46cccba4a57e092bf108695e8e88996f

Gate: get full process list from host

This PS adds the ps command to the gate log collection task.

Change-Id: Ife1710854547b28a6dfa6b3e4d527ebf4d04179b

Provide ability to specify proxy

This patch set provides ability to configure the proxy setting in
the ansible playbook used to run the gate.

Change-Id: I7ae66cb16fa6db7d46a0a3f23c289cc625e583c4
Signed-off-by: Tin Lam <tin@irrational.io>

Revert Elasticsearch/Kibana image change

This reverts the changes made to Elasticsearch, Kibana and fluent
logging charts in https://review.openstack.org/#/c/550229/7.

Specifically, this moves the images back to previous used versions
and makes the required changes to the fluent-logging elasticsearch
template job to include the correct mapping directives for the
elasticsearch template.

This change was made to give more time for evaluating a more
robust solution for switching to the official upstream images that
will not cause intermittent gate failures as seen since 550229 was
merged

Change-Id: I9f70b3412a8edc5cb1d80937b158aa2fe7b1ec82

Correcting a spelling error

perfoming -> performing

Change-Id: I639d8a4d0021dbdcd66ff711f86c25f1313a823f

Fluentd: Change elasticsearch endpoint port and creds reference

Updates the fluent-logging chart to reference the elasticsearch
endpoint via lookups on the `http` port to match the elasticsearch
chart's handling of the client port.  This also updates the helm
test pod to reference the elasticsearch credentials via the
secret used elsewhere in the fluent-logging chart

Change-Id: I352d912db5e231e14dc58cdf897ae642f3256373

Calico: update chart to deploy calicoctl via init-container

This PS moves the installation of calicoctl to an init-container.

Change-Id: I2dbd4c605889d86c87455d43a500d8a2697c49af

calico: only specify ipv6 listen address if we have one

Change-Id: I2bb381c227b06f5c511497b3e1720f9336f6d1c8

Support pod dependencies

Adds support for a new feature of kubernetes-entrypoint, pod
dependencies, that was added in v0.3.0.

Change-Id: I78d9e0545ca3b837cd2386783386a253f7f5a2d6

Nagios Chart

This adds the nagios chart to osh-infra to provide additional
monitoring functionality. It uses helper functions to consume
yaml definitions for services, commands, hosts and hostgroups
to generate the required configurations for those entities in
nagios's configuration

Change-Id: I6238bb8cb1e5c8dc48594ddea50693f3e7b0a176

add additional collectors to process exporter

Change-Id: I2c1055103a89663920cb309c714658caa3a74e05

Dynamically generate list of rules files for prometheus

This enables the dynamic generation of the list of rules files for
prometheus, driven by the rules added in the appropriate tree under
.Values.conf.prometheus.rules. This removes the necessity of adding
the file name manually in addition to defining the rules in the
rules tree, which should reduce overhead associated with adding
new rules for prometheus to evaluate

Change-Id: Ib768a252c5ea4f2d099df534c3ffcfb2949d7481

Prometheus: Generate command line flags dynamically

This proposes a means for generating the command line flags for
configuring the Prometheus service via the values file instead of
templating out the command line flags used for the service. This
allows flexibility in choosing which flags and values to use when
deploying Prometheus, without needing to modify the chart itself

Change-Id: I74845b96e213403ad743724137a82ce2c78fcd1f

Grafana dashboard for Nginx using nginx-vts metrics

Change-Id: Ifd0eeab3fcb2740ab53ce23ba1bcd2146f351aed

Helm-Toolkit: Reduce delta between OSH and OSH-Infra

This PS reduces the delta between OSH and OSH-Infra helm toolkits.

Change-Id: I00a684b3801a0990550f55c8facb0252fddb67f2

Helm-Toolkit: Reduce delta between OSH and OSH-Infra to image repo

This PS reduces the delta between OSH and OSH-Infra helm toolkits
to simply the image repo management functions.

Change-Id: I3addfbcda6a3c1d34c9f967be07eaceffb468f0e

Nagios: Configure ports with endpoint port lookups

This ps updates the nagios chart to use endpoint port lookups for
port configuration, bringing it in line with the other charts

Change-Id: I500b4741d50132f6c316ded660981e2af8b71e7a

Calico: Remove duplicate calico_cni image tag

This removes a duplicate calico_cni image tag from the calico
chart in osh-infra

Change-Id: I702796cf43ead9c10558a76bf4c23e133b0e8228

Update to F27 and update docker packages

This PS moves the fedora gates to use F27. It also updates
the docker packages used for both fedora and centos

Change-Id: Ic5ab5793d7176a81a8ea70c82b1d53a1fdf5f740

Move roles and playbooks to top level

This proposes moving the playbooks and roles to the top level of
osh-infra to be consumed by osh

Change-Id: I7e3516da8e516060f94b8b9c8875918504de7446

Helm-Toolkit: Kubernetes Entrypoint, simplify image dependencies

This PS simplify the logic for dyanmicly merging the image management
depenencies into pod deps when active.

Change-Id: I0cf6c93173bc5fbce697ac15be8697d3b1326d0a

OSH-infra: move charts to use ingress manifest in htk

This moves all relevant charts in osh-infra to use the htk manifest
template for ingresses, bringing them in line with the charts in
openstack-helm

Change-Id: Ic9c3cc6f0051fa66b6f88ec2b2725698b36ce824